requirements.txt
djangorestframework==3.16.0
black==25.1.0